Move add to wishlist to merlins connector
