+++ web/app/integration-manager/_merlins-connector/selectors.js
@@ -1,3 +1,4 @@
+import Immutable from 'immutable'
 import {createGetSelector} from 'reselect-immutable-helpers'
 
 export const getIntegrationManager = ({integrationManager}) => integrationManager
@@ -5,3 +6,5 @@ export const getCustomerEntityID = createGetSelector(getIntegrationManager, 'cus
 export const getFormKey = createGetSelector(getIntegrationManager, 'formKey')
 export const getLoginHref = createGetSelector(getIntegrationManager, 'loginHref')
 export const getRegisterHref = createGetSelector(getIntegrationManager, 'registerHref')
+export const getIMProductDetailsByPathKey = (pathKey) => createGetSelector(getIntegrationManager, pathKey, Immutable.Map())
+export const getUenc = (pathKey) => createGetSelector(getIMProductDetailsByPathKey(pathKey), 'uenc')
