+++ web/app/integration-manager/cart/commands.js
@@ -11,3 +11,5 @@ export const fetchCartPageData = (...args) => connector.fetchCartPageData(...arg
 export const removeFromCart = (...args) => connector.removeFromCart(...args)
 
 export const updateItemQuantity = (...args) => connector.updateItemQuantity(...args)
+
+export const addToWishlist = (...args) => connector.addToWishlist(...args)
